docs: rewrite readme with full guide
- add quick start and architecture sections
- add contract test config and fixtures
- update dev dependencies
- regenerate api docs to match new src layout
