+++ .codegraph/daemon.pid
@@ -1,0 +1,6 @@
+{
+  "pid": 2939120,
+  "version": "0.9.9",
+  "socketPath": "/home/milos/code/msh/packages/app-boot/.codegraph/daemon.sock",
+  "startedAt": 1782739666217
+}
