+++ package.json
@@ -69,6 +69,7 @@
 		"test-coverage-unit-publish": "codecov",
 		"test:e2e": "cd test && npm run test",
 		"test:int": "vitest --config=./vitest.config.int.ts",
+		"test:contract": "vitest --config=./vitest.config.contract.ts",
 		"test:unit": "vitest --config=./vitest.config.unit.ts",
 		"tsc": "tsc -p tsconfig.build.json",
 		"tsc-check": "npm run tsc -- --noEmit",
@@ -91,24 +92,24 @@
 		"@beecode/msh-logger": "^2.0.0"
 	},
 	"devDependencies": {
-		"@beecode/msh-test-contractor": "^0.4.2",
-		"@commitlint/cli": "^21.0.2",
-		"@commitlint/config-conventional": "^21.0.2",
-		"@commitlint/prompt": "^21.0.2",
-		"@types/node": "^25.9.3",
-		"@vitest/coverage-v8": "^4.1.8",
+		"@beecode/msh-test-contractor": "^1.0.0",
+		"@commitlint/cli": "^21.2.0",
+		"@commitlint/config-conventional": "^21.2.0",
+		"@commitlint/prompt": "^21.2.0",
+		"@types/node": "^26.0.1",
+		"@vitest/coverage-v8": "^4.1.9",
 		"commitizen": "^4.3.2",
 		"concurrently": "^10.0.3",
 		"husky": "^9.1.7",
 		"json-sort-cli": "^4.1.3",
-		"lint-staged": "^17.0.7",
+		"lint-staged": "^17.0.8",
 		"markdown-toc": "^1.2.0",
 		"rimraf": "^6.1.3",
-		"tsc-watch": "^7.2.0",
+		"tsc-watch": "^7.2.1",
 		"typedoc": "^0.28.19",
 		"typedoc-plugin-markdown": "^4.12.0",
 		"typescript": "^6.0.3",
-		"vitest": "^4.1.8"
+		"vitest": "^4.1.9"
 	},
 	"engines": {
 		"node": ">=20.8.1",
