--- resource/doc/api/.nojekyll
@@ -1,1 +1,0 @@
-TypeDoc added this file to prevent GitHub Pages from using Jekyll. You can turn off this behavior by setting the `githubPages` option to false.
