+++ resource/doc/api/business/service/app-flow/README.md
@@ -1,0 +1,19 @@
+[**@beecode/msh-app-boot**](../../../README.md)
+
+***
+
+[@beecode/msh-app-boot](../../../README.md) / business/service/app-flow
+
+# business/service/app-flow
+
+## Enumerations
+
+- [FlowDirectionMapper](enumerations/FlowDirectionMapper.md)
+
+## Classes
+
+- [AppFlow](classes/AppFlow.md)
+
+## Type Aliases
+
+- [FlowList](type-aliases/FlowList.md)
