+++ resource/doc/api/business/service/app-flow/enumerations/FlowDirectionMapper.md
@@ -1,0 +1,25 @@
+[**@beecode/msh-app-boot**](../../../../README.md)
+
+***
+
+[@beecode/msh-app-boot](../../../../README.md) / [business/service/app-flow](../README.md) / FlowDirectionMapper
+
+# Enumeration: FlowDirectionMapper
+
+Defined in: [business/service/app-flow.ts:6](https://github.com/beecode-rs/msh-app-boot/blob/f03720bc0da2ffbc23eb0b04dd0c7dcf87ac181d/src/business/service/app-flow.ts#L6)
+
+## Enumeration Members
+
+### CREATE
+
+> **CREATE**: `"create"`
+
+Defined in: [business/service/app-flow.ts:8](https://github.com/beecode-rs/msh-app-boot/blob/f03720bc0da2ffbc23eb0b04dd0c7dcf87ac181d/src/business/service/app-flow.ts#L8)
+
+***
+
+### DESTROY
+
+> **DESTROY**: `"destroy"`
+
+Defined in: [business/service/app-flow.ts:7](https://github.com/beecode-rs/msh-app-boot/blob/f03720bc0da2ffbc23eb0b04dd0c7dcf87ac181d/src/business/service/app-flow.ts#L7)
