+++ resource/doc/api/business/service/app-flow/type-aliases/FlowList.md
@@ -1,0 +1,11 @@
+[**@beecode/msh-app-boot**](../../../../README.md)
+
+***
+
+[@beecode/msh-app-boot](../../../../README.md) / [business/service/app-flow](../README.md) / FlowList
+
+# Type Alias: FlowList
+
+> **FlowList** = ([`LifeCycle`](../../life-cycle/classes/LifeCycle.md) \| [`LifeCycle`](../../life-cycle/classes/LifeCycle.md)[])[]
+
+Defined in: [business/service/app-flow.ts:4](https://github.com/beecode-rs/msh-app-boot/blob/f03720bc0da2ffbc23eb0b04dd0c7dcf87ac181d/src/business/service/app-flow.ts#L4)
