+++ resource/doc/api/business/service/life-cycle/README.md
@@ -1,0 +1,11 @@
+[**@beecode/msh-app-boot**](../../../README.md)
+
+***
+
+[@beecode/msh-app-boot](../../../README.md) / business/service/life-cycle
+
+# business/service/life-cycle
+
+## Classes
+
+- [LifeCycle](classes/LifeCycle.md)
