+++ resource/doc/api/controller/app-starter/README.md
@@ -1,0 +1,15 @@
+[**@beecode/msh-app-boot**](../../README.md)
+
+***
+
+[@beecode/msh-app-boot](../../README.md) / controller/app-starter
+
+# controller/app-starter
+
+## Enumerations
+
+- [AppStarterStatusMapper](enumerations/AppStarterStatusMapper.md)
+
+## Classes
+
+- [AppStarter](classes/AppStarter.md)
