+++ resource/doc/api/controller/app-starter/enumerations/AppStarterStatusMapper.md
@@ -1,0 +1,25 @@
+[**@beecode/msh-app-boot**](../../../README.md)
+
+***
+
+[@beecode/msh-app-boot](../../../README.md) / [controller/app-starter](../README.md) / AppStarterStatusMapper
+
+# Enumeration: AppStarterStatusMapper
+
+Defined in: [controller/app-starter.ts:4](https://github.com/beecode-rs/msh-app-boot/blob/f03720bc0da2ffbc23eb0b04dd0c7dcf87ac181d/src/controller/app-starter.ts#L4)
+
+## Enumeration Members
+
+### STARTED
+
+> **STARTED**: `"started"`
+
+Defined in: [controller/app-starter.ts:5](https://github.com/beecode-rs/msh-app-boot/blob/f03720bc0da2ffbc23eb0b04dd0c7dcf87ac181d/src/controller/app-starter.ts#L5)
+
+***
+
+### STOPPED
+
+> **STOPPED**: `"stopped"`
+
+Defined in: [controller/app-starter.ts:6](https://github.com/beecode-rs/msh-app-boot/blob/f03720bc0da2ffbc23eb0b04dd0c7dcf87ac181d/src/controller/app-starter.ts#L6)
