--- resource/doc/api/enums/app_flow.FlowDirectionMapper.md
@@ -1,32 +1,0 @@
-[@beecode/msh-app-boot](../README.md) / [app-flow](../modules/app_flow.md) / FlowDirectionMapper
-
-# Enumeration: FlowDirectionMapper
-
-[app-flow](../modules/app_flow.md).FlowDirectionMapper
-
-## Table of contents
-
-### Enumeration Members
-
-- [CREATE](app_flow.FlowDirectionMapper.md#create)
-- [DESTROY](app_flow.FlowDirectionMapper.md#destroy)
-
-## Enumeration Members
-
-### CREATE
-
-• **CREATE** = ``"create"``
-
-#### Defined in
-
-[app-flow.ts:8](https://github.com/beecode-rs/msh-app-boot/blob/ff89a8e/src/app-flow.ts#L8)
-
-___
-
-### DESTROY
-
-• **DESTROY** = ``"destroy"``
-
-#### Defined in
-
-[app-flow.ts:7](https://github.com/beecode-rs/msh-app-boot/blob/ff89a8e/src/app-flow.ts#L7)
