--- resource/doc/api/enums/app_starter.AppStarterStatusMapper.md
@@ -1,32 +1,0 @@
-[@beecode/msh-app-boot](../README.md) / [app-starter](../modules/app_starter.md) / AppStarterStatusMapper
-
-# Enumeration: AppStarterStatusMapper
-
-[app-starter](../modules/app_starter.md).AppStarterStatusMapper
-
-## Table of contents
-
-### Enumeration Members
-
-- [STARTED](app_starter.AppStarterStatusMapper.md#started)
-- [STOPPED](app_starter.AppStarterStatusMapper.md#stopped)
-
-## Enumeration Members
-
-### STARTED
-
-• **STARTED** = ``"started"``
-
-#### Defined in
-
-[app-starter.ts:5](https://github.com/beecode-rs/msh-app-boot/blob/ff89a8e/src/app-starter.ts#L5)
-
-___
-
-### STOPPED
-
-• **STOPPED** = ``"stopped"``
-
-#### Defined in
-
-[app-starter.ts:6](https://github.com/beecode-rs/msh-app-boot/blob/ff89a8e/src/app-starter.ts#L6)
