+++ resource/doc/api/index/README.md
@@ -1,0 +1,49 @@
+[**@beecode/msh-app-boot**](../README.md)
+
+***
+
+[@beecode/msh-app-boot](../README.md) / index
+
+# index
+
+## References
+
+### AppFlow
+
+Re-exports [AppFlow](../business/service/app-flow/classes/AppFlow.md)
+
+***
+
+### AppStarter
+
+Re-exports [AppStarter](../controller/app-starter/classes/AppStarter.md)
+
+***
+
+### AppStarterStatusMapper
+
+Re-exports [AppStarterStatusMapper](../controller/app-starter/enumerations/AppStarterStatusMapper.md)
+
+***
+
+### FlowDirectionMapper
+
+Re-exports [FlowDirectionMapper](../business/service/app-flow/enumerations/FlowDirectionMapper.md)
+
+***
+
+### FlowList
+
+Re-exports [FlowList](../business/service/app-flow/type-aliases/FlowList.md)
+
+***
+
+### LifeCycle
+
+Re-exports [LifeCycle](../business/service/life-cycle/classes/LifeCycle.md)
+
+***
+
+### setAppBootLogger
+
+Re-exports [setAppBootLogger](../util/logger/functions/setAppBootLogger.md)
