--- resource/doc/api/modules/app_flow.md
@@ -1,27 +1,0 @@
-[@beecode/msh-app-boot](../README.md) / app-flow
-
-# Module: app-flow
-
-## Table of contents
-
-### Enumerations
-
-- [FlowDirectionMapper](../enums/app_flow.FlowDirectionMapper.md)
-
-### Classes
-
-- [AppFlow](../classes/app_flow.AppFlow.md)
-
-### Type Aliases
-
-- [FlowList](app_flow.md#flowlist)
-
-## Type Aliases
-
-### FlowList
-
-Ƭ **FlowList**: ([`LifeCycle`](../classes/life_cycle.LifeCycle.md) \| [`LifeCycle`](../classes/life_cycle.LifeCycle.md)[])[]
-
-#### Defined in
-
-[app-flow.ts:4](https://github.com/beecode-rs/msh-app-boot/blob/ff89a8e/src/app-flow.ts#L4)
