--- resource/doc/api/modules/app_starter.md
@@ -1,13 +1,0 @@
-[@beecode/msh-app-boot](../README.md) / app-starter
-
-# Module: app-starter
-
-## Table of contents
-
-### Enumerations
-
-- [AppStarterStatusMapper](../enums/app_starter.AppStarterStatusMapper.md)
-
-### Classes
-
-- [AppStarter](../classes/app_starter.AppStarter.md)
