--- resource/doc/api/modules/index.md
@@ -1,50 +1,0 @@
-[@beecode/msh-app-boot](../README.md) / index
-
-# Module: index
-
-## Table of contents
-
-### References
-
-- [AppFlow](index.md#appflow)
-- [AppStarter](index.md#appstarter)
-- [AppStarterStatusMapper](index.md#appstarterstatusmapper)
-- [FlowDirectionMapper](index.md#flowdirectionmapper)
-- [FlowList](index.md#flowlist)
-- [LifeCycle](index.md#lifecycle)
-
-## References
-
-### AppFlow
-
-Re-exports [AppFlow](../classes/app_flow.AppFlow.md)
-
-___
-
-### AppStarter
-
-Re-exports [AppStarter](../classes/app_starter.AppStarter.md)
-
-___
-
-### AppStarterStatusMapper
-
-Re-exports [AppStarterStatusMapper](../enums/app_starter.AppStarterStatusMapper.md)
-
-___
-
-### FlowDirectionMapper
-
-Re-exports [FlowDirectionMapper](../enums/app_flow.FlowDirectionMapper.md)
-
-___
-
-### FlowList
-
-Re-exports [FlowList](app_flow.md#flowlist)
-
-___
-
-### LifeCycle
-
-Re-exports [LifeCycle](../classes/life_cycle.LifeCycle.md)
