--- resource/doc/api/modules/index_jest_setup.md
@@ -1,3 +1,0 @@
-[@beecode/msh-app-boot](../README.md) / index-jest-setup
-
-# Module: index-jest-setup
