--- resource/doc/api/modules/life_cycle.md
@@ -1,9 +1,0 @@
-[@beecode/msh-app-boot](../README.md) / life-cycle
-
-# Module: life-cycle
-
-## Table of contents
-
-### Classes
-
-- [LifeCycle](../classes/life_cycle.LifeCycle.md)
