--- resource/doc/api/modules/util_logger.md
@@ -1,44 +1,0 @@
-[@beecode/msh-app-boot](../README.md) / util/logger
-
-# Module: util/logger
-
-## Table of contents
-
-### Functions
-
-- [logger](util_logger.md#logger)
-- [setAppBootLogger](util_logger.md#setappbootlogger)
-
-## Functions
-
-### logger
-
-▸ **logger**(): `LoggerStrategy`
-
-#### Returns
-
-`LoggerStrategy`
-
-#### Defined in
-
-[util/logger.ts:12](https://github.com/beecode-rs/msh-app-boot/blob/ff89a8e/src/util/logger.ts#L12)
-
-___
-
-### setAppBootLogger
-
-▸ **setAppBootLogger**(`logger`): `void`
-
-#### Parameters
-
-| Name | Type |
-| :------ | :------ |
-| `logger` | `LoggerStrategy` |
-
-#### Returns
-
-`void`
-
-#### Defined in
-
-[util/logger.ts:8](https://github.com/beecode-rs/msh-app-boot/blob/ff89a8e/src/util/logger.ts#L8)
