+++ resource/doc/api/util/logger/README.md
@@ -1,0 +1,12 @@
+[**@beecode/msh-app-boot**](../../README.md)
+
+***
+
+[@beecode/msh-app-boot](../../README.md) / util/logger
+
+# util/logger
+
+## Functions
+
+- [logger](functions/logger.md)
+- [setAppBootLogger](functions/setAppBootLogger.md)
