+++ resource/doc/api/util/logger/functions/logger.md
@@ -1,0 +1,15 @@
+[**@beecode/msh-app-boot**](../../../README.md)
+
+***
+
+[@beecode/msh-app-boot](../../../README.md) / [util/logger](../README.md) / logger
+
+# Function: logger()
+
+> **logger**(): `LoggerStrategy`
+
+Defined in: [util/logger.ts:12](https://github.com/beecode-rs/msh-app-boot/blob/f03720bc0da2ffbc23eb0b04dd0c7dcf87ac181d/src/util/logger.ts#L12)
+
+## Returns
+
+`LoggerStrategy`
