+++ resource/doc/api/util/logger/functions/setAppBootLogger.md
@@ -1,0 +1,21 @@
+[**@beecode/msh-app-boot**](../../../README.md)
+
+***
+
+[@beecode/msh-app-boot](../../../README.md) / [util/logger](../README.md) / setAppBootLogger
+
+# Function: setAppBootLogger()
+
+> **setAppBootLogger**(`logger`): `void`
+
+Defined in: [util/logger.ts:8](https://github.com/beecode-rs/msh-app-boot/blob/f03720bc0da2ffbc23eb0b04dd0c7dcf87ac181d/src/util/logger.ts#L8)
+
+## Parameters
+
+### logger
+
+`LoggerStrategy`
+
+## Returns
+
+`void`
