--- src/__mocks__/life-cycle-mock-implementation.ts
@@ -1,22 +1,0 @@
-import { type Mock, vi } from 'vitest'
-
-import { LifeCycle } from '#src/life-cycle.js'
-
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-export class LifeCycleMockImplementation<T = any> extends LifeCycle<T> {
-	// eslint-disable-next-line @typescript-eslint/no-useless-constructor
-	constructor(params: { name: string }) {
-		super(params)
-	}
-
-	protected _createFn = vi.fn()
-	protected _destroyFn = vi.fn()
-
-	get createFn(): Mock {
-		return this._createFn
-	}
-
-	get destroyFn(): Mock {
-		return this._destroyFn
-	}
-}
