+++ src/index.ts
@@ -1,6 +1,8 @@
 /* istanbul ignore file */
-export { LifeCycle } from '#src/life-cycle.js'
+export { LifeCycle } from '#src/business/service/life-cycle.js'
 
-export { AppFlow, FlowList, FlowDirectionMapper } from '#src/app-flow.js'
+export { AppFlow, FlowList, FlowDirectionMapper } from '#src/business/service/app-flow.js'
 
-export { AppStarter, AppStarterStatusMapper } from '#src/app-starter.js'
+export { AppStarter, AppStarterStatusMapper } from '#src/controller/app-starter.js'
+
+export { setAppBootLogger } from '#src/util/logger.js'
