+++ typedoc.json
@@ -5,7 +5,11 @@
 	],
 	"exclude": [
 		"./node_modules/**/*",
-		"./**/*.test.ts"
+		"./**/*.test.ts",
+		"./**/*.contract.ts",
+		"./**/__mocks__/**/*",
+		"./**/__tests__/**/*",
+		"./**/__fixtures__/**/*"
 	],
 	"json": "./resource/doc/api/docs.json",
 	"out": "./resource/doc/api",
